requirements.txt
azure-identity
azure-keyvault
azure-keyvault-secrets
azure-storage-file-datalake
joblib
numpy
pandas
python-dotenv